requirements.txt
absl-py==0.10.0
argon2-cffi==20.1.0
asgiref==3.2.8
astunparse==1.6.3
attrs==20.1.0
backcall==0.1.0
beautifulsoup4==4.9.1
bleach==3.1.5
branca @ git+https://github.com/python-visualization/branca.git@5887b9bce65be873c9c2ad867b3d6334b672b780
cachetools==4.1.1
certifi==2020.6.20
cffi==1.14.2
chardet==3.0.4
click==7.1.2
click-plugins==1.1.1
cligj==0.5.0
colorama==0.4.3
coreapi==2.3.3
coreschema==0.0.4
cssselect==1.1.0
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
dj-database-url==0.5.0
Django==3.0.7
django-allauth==0.42.0
django-bootstrap4==2.0.1
django-cors-headers==3.3.0
django-extensions==2.2.9
django-rest-auth==0.9.5
django-rest-swagger==2.0.7
djangorestframework==3.11.0
entrypoints==0.3
et-xmlfile==1.0.1
Faker==4.1.0
folium==0.11.0
gast==0.3.3
gitdb==4.0.5
GitPython==3.1.8
google-auth==1.21.0
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio==1.31.0
gunicorn==20.0.4
h5py==2.10.0
idna==2.9
ipykernel==5.3.4
ipython==7.15.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
itypes==1.2.0
jdcal==1.4.1
jedi==0.17.0
Jinja2==2.11.2
joblib==0.16.0
json5==0.9.5
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.7
jupyter-console==6.2.0
jupyter-contrib-core==0.3.3
jupyter-contrib-nbextensions==0.5.1
jupyter-core==4.6.3
jupyter-highlight-selected-word==0.2.0
jupyter-latex-envs==1.4.6
jupyter-nbextensions-configurator==0.4.1
jupyterlab==2.2.6
jupyterlab-git==0.21.1
jupyterlab-server==1.2.0
jupyterthemes==0.20.0
Keras-Preprocessing==1.1.2
kiwisolver==1.2.0
lesscpy==0.14.0
lxml==4.5.2
Markdown==3.2.2
MarkupSafe==1.1.1
matplotlib==3.3.2
missingno==0.4.2
mistune==0.8.4
MouseInfo==0.1.3
munch==2.5.0
nbconvert==5.6.1
nbdime==2.0.0
nbformat==5.0.7
node==0.9.25
notebook==6.1.3
numpy==1.18.5
oauthlib==3.1.0
odict==1.7.0
openapi-codec==1.3.2
openpyxl==3.0.5
opt-einsum==3.3.0
packaging==20.4
pandas==1.1.1
pandocfilters==1.4.2
parso==0.7.0
patsy==0.5.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.2.0
plumber==1.5
ply==3.11
prometheus-client==0.8.0
prompt-toolkit==3.0.5
protobuf==3.13.0
psycopg2-binary==2.8.6
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAutoGUI==0.9.50
pycparser==2.20
PyGetWindow==0.0.8
Pygments==2.6.1
PyMsgBox==1.0.8
pyparsing==2.4.7
pyperclip==1.8.0
pyproj==2.6.1.post1
pyquery==1.4.1
PyRect==0.1.4
pyrsistent==0.16.0
PyScreeze==0.1.26
python-dateutil==2.8.1
python-decouple==3.3
python3-openid==3.1.0
PyTweening==1.0.3
pytz==2020.1
pywin32==228
pywinpty==0.5.7
PyYAML==5.3.1
pyzmq==19.0.2
qtconsole==4.7.7
QtPy==1.9.0
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.6
scikit-learn==0.23.2
scipy==1.4.1
seaborn==0.10.1
selenium==3.141.0
Send2Trash==1.5.0
Shapely==1.7.1
simplejson==3.17.2
six==1.9.0
sklearn==0.0
smmap==3.0.4
soupsieve==2.0.1
sqlparse==0.3.1
statsmodels==0.12.0
tensorboard==2.3.0
tensorboard-plugin-wit==1.7.0
tensorflow==2.3.0
tensorflow-estimator==2.3.0
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
text-unidecode==1.3
threadpoolctl==2.1.0
tornado==6.0.4
traitlets==4.3.3
uritemplate==3.0.1
urllib3==1.25.9
wcwidth==0.2.4
webencodings==0.5.1
Werkzeug==1.0.1
whitenoise==5.2.0
widgetsnbextension==3.5.1
wrapt==1.12.1
xlrd==1.2.0
zope.component==4.6.1
zope.deferredimport==4.3.1
zope.deprecation==4.4.0
zope.event==4.4
zope.hookable==5.0.1
zope.interface==5.1.0
zope.lifecycleevent==4.3
zope.proxy==4.3.5
